# backend/docs/openapi.yaml
definitions:
...
  models.News:
    properties:
      category:
        description: Category or topic
        example: startup
        type: string
      created_at:
        description: Creation timestamp (UTC)
        format: date-time
        type: string
      description:
        description: Body text
        example: Series A raised
        type: string
      id:
        description: Unique news identifier
        example: 1
        type: integer
      image_url:
        description: Image URL
        example: https://cdn.example.com/news/1.png
        format: uri
        type: string
      location:
        description: Location (if relevant)
        example: Paris
        type: string
      news_date:
        description: Publication date
        format: date
        type: string
      startup_id:
        description: Related startup ID
        example: 1
        type: integer
      title:
        description: Title
        example: Funding round
        type: string
      updated_at:
        description: Update timestamp (UTC)
        format: date-time
        type: string
    type: object
...
  requests.NewsCreateRequest:
    properties:
      category:
        description: Category or topic
        example: startup
        type: string
      description:
        description: Body text
        example: Series A raised
        type: string
      image_url:
        description: Image URL
        example: https://cdn.example.com/news/1.png
        format: uri
        type: string
      location:
        description: Location (if relevant)
        example: Paris
        type: string
      news_date:
        description: Publication date (YYYY-MM-DD)
        example: "2025-09-01"
        format: date
        type: string
      startup_id:
        description: Related startup ID
        example: 1
        type: integer
      title:
        description: News title
        example: Funding round
        type: string
    type: object
  requests.NewsUpdateRequest:
    properties:
      category:
        description: New category
        example: event
        type: string
      description:
        description: New body text
        example: Updated description
        type: string
      image_url:
        description: New image URL
        example: https://cdn.example.com/news/2.png
        format: uri
        type: string
      location:
        description: New location
        example: Lyon
        type: string
      news_date:
        description: New publication date
        example: "2025-09-02"
        format: date
        type: string
      startup_id:
        description: New related startup
        example: 2
        type: integer
      title:
        description: New title
        example: Updated title
        type: string
    type: object
...
  response.ErrorBody:
    properties:
      code:
        description: Machine-readable error code
        enum:
        - invalid_params
        - invalid_sort
        - invalid_payload
        - internal_error
        - not_found
        - email_taken
        - invalid_token
        - invalid_credentials
        - no_fields
        - method_not_allowed
        - unauthorized
        example: internal_error
        type: string
      details:
        description: Optional granular details (validation errors, etc.)
      message:
        description: Human-readable error message
        example: failed to retrieve resource
        type: string
    type: object
...
  response.ImageURLResponse:
    properties:
      image_url:
        type: string
    type: object
...
  response.MessageResponse:
    properties:
      message:
        type: string
    type: object
  response.NewsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.News'
        type: array
      pagination:
        $ref: '#/definitions/response.PageMeta'
    type: object
  response.NewsObjectResponse:
    properties:
      data:
        $ref: '#/definitions/models.News'
    type: object
...
  response.PageMeta:
    properties:
      has_next:
        description: Whether there is a next page
        example: true
        type: boolean
      has_prev:
        description: Whether there is a previous page
        example: false
        type: boolean
      page:
        description: Current page number (1-based)
        example: 1
        minimum: 1
        type: integer
      per_page:
        description: Page size (items per page)
        example: 20
        minimum: 1
        type: integer
      total:
        description: Total number of items available
        example: 123
        minimum: 0
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /admin/news:
    post:
      consumes:
      - application/json
      description: Creates a news item (admin required).
      parameters:
      - description: News
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.NewsCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.NewsObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create news
      tags:
      - News
  /admin/news/{id}:
    delete:
      description: Deletes a news item (admin required).
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete news
      tags:
      - News
    patch:
      consumes:
      - application/json
      description: Updates a news item (admin required).
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.NewsUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NewsObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update news
      tags:
      - News
...
  /news:
    get:
      description: Returns a paginated list of news with sorting.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: per_page
        type: integer
      - default: news_date
        description: Sort field
        enum:
        - id
        - title
        - news_date
        - category
        - startup_id
        - created_at
        - updated_at
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NewsListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: List news
      tags:
      - News
  /news/{id}:
    get:
      description: Retrieves a news item by ID.
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.NewsObjectResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Get news item
      tags:
      - News
  /news/{id}/image:
    get:
      description: Returns the image URL of a news item.
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ImageURLResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: News image URL
      tags:
      - News
...
swagger: "2.0"